Don't skip type check of value argument when it's known
E.g., in case when type arguments are explicitly specified.

The problem appears when resolving recursive calls (see sameTypeParameterUse.kt):

`foo<R>(x, "")` call after substitution has R as type of it's first argument,
and when checking value arguments, x is not checked because supposed
to be dependent on type parameter of candidate (that is effectively false),
so there are two successful candidates in resolution results and
wrong OVERLOAD_RESOLUTION_AMBIGUITY error.

Solution is just to force type check for such cases with explicit type parameters given.